NGSTACK-460 enable deeper levels of menu items
