+++ bundle/Menu/Factory/LocationFactory/MenuItemExtension.php
@@ -4,15 +4,10 @@ declare(strict_types=1);
 
 namespace Netgen\Bundle\SiteBundle\Menu\Factory\LocationFactory;
 
-use eZ\Publish\API\Repository\Values\Content\LocationQuery;
-use eZ\Publish\API\Repository\Values\Content\Query\Criterion;
-use eZ\Publish\API\Repository\Values\Content\Search\SearchHit;
 use eZ\Publish\Core\FieldType\Url\Value as UrlValue;
 use eZ\Publish\Core\MVC\ConfigResolverInterface;
 use eZ\Publish\Core\MVC\Symfony\SiteAccess\URILexer;
 use Knp\Menu\ItemInterface;
-use Netgen\EzPlatformSiteApi\API\FilterService;
-use Netgen\EzPlatformSiteApi\API\LoadService;
 use Netgen\EzPlatformSiteApi\API\Values\Content;
 use Netgen\EzPlatformSiteApi\API\Values\Location;
 use Psr\Log\LoggerInterface;
@@ -20,24 +15,12 @@ use Psr\Log\NullLogger;
 use Symfony\Cmf\Component\Routing\RouteObjectInterface;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
-use Throwable;
-use function array_map;
 use function in_array;
 use function mb_stripos;
 use function sprintf;
 
 class MenuItemExtension implements ExtensionInterface
 {
-    /**
-     * @var \Netgen\EzPlatformSiteApi\API\LoadService
-     */
-    protected $loadService;
-
-    /**
-     * @var \Netgen\EzPlatformSiteApi\API\FilterService
-     */
-    protected $filterService;
-
     /**
      * @var \Symfony\Component\Routing\Generator\UrlGeneratorInterface
      */
@@ -58,19 +41,22 @@ class MenuItemExtension implements ExtensionInterface
      */
     protected $logger;
 
+    /**
+     * @var \Netgen\Bundle\SiteBundle\Menu\Factory\LocationFactory\ChildrenBuilder
+     */
+    protected $childrenBuilder;
+
     public function __construct(
-        LoadService $loadService,
-        FilterService $filterService,
         UrlGeneratorInterface $urlGenerator,
         RequestStack $requestStack,
         ConfigResolverInterface $configResolver,
+        ChildrenBuilder $childrenBuilder,
         ?LoggerInterface $logger = null
     ) {
-        $this->loadService = $loadService;
-        $this->filterService = $filterService;
         $this->urlGenerator = $urlGenerator;
         $this->requestStack = $requestStack;
         $this->configResolver = $configResolver;
+        $this->childrenBuilder = $childrenBuilder;
         $this->logger = $logger ?? new NullLogger();
     }
 
@@ -88,7 +74,7 @@ class MenuItemExtension implements ExtensionInterface
                 ->setLinkAttribute('rel', 'nofollow noopener noreferrer');
         }
 
-        $this->buildChildItems($item, $location->content);
+        $this->childrenBuilder->buildChildItems($item, $location->content);
     }
 
     protected function buildItemFromContent(ItemInterface $item, Content $content): void
@@ -156,86 +142,6 @@ class MenuItemExtension implements ExtensionInterface
             // Disable link for content types that act as simple content containers
             // and that have no their own full views
             $item->setUri('');
-        }
-    }
-
-    protected function buildChildItems(ItemInterface $item, Content $content): void
-    {
-        $childLocations = [];
-
-        if (!$content->getField('parent_node')->isEmpty()) {
-            $destinationContent = $content->getFieldRelation('parent_node');
-            if (!$destinationContent instanceof Content) {
-                return;
-            }
-
-            if (!$destinationContent->mainLocation->isVisible) {
-                $this->logger->error(sprintf('Menu item (#%s) has a related object (#%s) that is not visible.', $content->id, $destinationContent->id));
-
-                return;
-            }
-
-            $criteria = [
-                new Criterion\Visibility(Criterion\Visibility::VISIBLE),
-                new Criterion\ParentLocationId($destinationContent->mainLocation->id),
-            ];
-
-            if (!$content->getField('class_filter')->isEmpty() && !$content->getField('class_filter_type')->isEmpty()) {
-                /** @var \Netgen\Bundle\ContentTypeListBundle\Core\FieldType\ContentTypeList\Value $contentTypeFilter */
-                $contentTypeFilter = $content->getField('class_filter')->value;
-
-                /** @var \Netgen\Bundle\EnhancedSelectionBundle\Core\FieldType\EnhancedSelection\Value $filterType */
-                $filterType = $content->getField('class_filter_type')->value;
-
-                if ($filterType->identifiers[0] === 'include') {
-                    $criteria[] = new Criterion\ContentTypeIdentifier($contentTypeFilter->identifiers);
-                } elseif ($filterType->identifiers[0] === 'exclude') {
-                    $criteria[] = new Criterion\LogicalNot(
-                        new Criterion\ContentTypeIdentifier($contentTypeFilter->identifiers)
-                    );
-                }
-            }
-
-            $query = new LocationQuery();
-            $query->filter = new Criterion\LogicalAnd($criteria);
-            $query->sortClauses = $destinationContent->mainLocation->innerLocation->getSortClauses();
-
-            if (!$content->getField('limit')->isEmpty()) {
-                /** @var \eZ\Publish\Core\FieldType\Integer\Value $limit */
-                $limit = $content->getField('limit')->value;
-                if ($limit->value > 0) {
-                    $query->limit = $limit->value;
-                }
-            }
-
-            $searchResult = $this->filterService->filterLocations($query);
-
-            $childLocations = array_map(
-                static function (SearchHit $searchHit) {
-                    return $searchHit->valueObject;
-                },
-                $searchResult->searchHits
-            );
-        } elseif (!$content->getField('menu_items')->isEmpty()) {
-            foreach ($content->getField('menu_items')->value->destinationContentIds as $destinationContentId) {
-                if (empty($destinationContentId)) {
-                    $this->logger->error(sprintf('Empty content ID in RelationList field "%s" for content #%s', 'menu_items', $content->id));
-
-                    continue;
-                }
-
-                try {
-                    $childLocations[] = $this->loadService->loadContent($destinationContentId)->mainLocation;
-                } catch (Throwable $t) {
-                    $this->logger->error($t->getMessage());
-
-                    continue;
-                }
-            }
-        }
-
-        foreach ($childLocations as $location) {
-            $item->addChild(null, ['ezlocation' => $location]);
         }
     }
 }
